VM-container-*: Added test for update process using nullos
Generate configs for nullos in version 1 to 3. Use version to indicate
1GB, 2GB, 3GB test images. Image files are raw images with all blocks
set to zero. The Guestos configs are generated similarity to container
test configs and signed on build host with recalculated values for the
image. The corresponding image files are generated on demand by 'dd'
using on "/var/volatile/tmp" in the running c0 instance.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>