vars/stepInitWs.groovy: check if meta-* repo exists before removing
If building with PR set for an meta-gyroidos-* repo, check if it
exists. In case the piple is triggered for diffrent machines, but
with a PR depending machine repo, this would fail for other machines
which do not check out this repo. E.g., one needs a depending PR
in meta-gyroidos-nxp but triggers the pipeline through the cml repo.
In this case the meta-gyroidos-nxp repo is not in the manifest.xml
of the workspace and the remove-project would fail. In that case just
continue the loop.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>